feat(ci): include zstd in the pongo runner image
for a number of plugins, we use the zstd lib which needs to be compiled.
The compile operation takes around 2 minutes, which when included
dynamically via pongo-setup scripts local to a plugin means every test
run takes two minutes just in startup time. Including the zstd lib in
the image means the setup time is only paide once on the initial pongo
build and tests do not need to run the compile steps